requirements.txt
neo4j==5.14.1
pandas>=2.0.3
python-dotenv==1.0.0
flask==3.0.0
werkzeug==3.0.1
huggingface-hub>=0.19.4,<0.21.0
torch>=2.6.0
flask-cors